+++ src/main/java/fuzs/puzzleslib/element/side/IServerElement.java
@@ -10,20 +10,37 @@ public interface IServerElement extends ISidedElement {
     /**
      * register server events
      */
+    default void constructServer() {
+
+        this.setupServer();
+    }
+
+    /**
+     * register server events
+     */
+    @Deprecated
     default void setupServer() {
 
     }
 
     /**
+     * TODO rename to #setupServer
      * setup for {@link net.minecraftforge.fml.event.lifecycle.FMLDedicatedServerSetupEvent}
      * is always loaded no matter the element's state
+     */
+    default void setupServer2() {
+
+    }
+
+    /**
+     * called when the element is enabled and on launch
      */
     default void loadServer() {
 
     }
 
     /**
-     * should basically clean up changes made by this element
+     * called when the element is disabled
      */
     default void unloadServer() {