+++ src/main/java/fuzs/puzzleslib/proxy/ServerProxy.java
@@ -12,8 +12,15 @@ import javax.annotation.Nonnull;
  */
 public class ServerProxy implements IProxy<MinecraftServer> {
 
+    /**
+     * private singleton constructor
+     */
+    private ServerProxy() {
+
+    }
+
     @Override
-    public MinecraftServer getInstance() {
+    public MinecraftServer getGameInstance() {
 
         return LogicalSidedProvider.INSTANCE.get(LogicalSide.SERVER);
     }
@@ -23,6 +30,24 @@ public class ServerProxy implements IProxy<MinecraftServer> {
     public PlayerEntity getPlayer(PlayerEntity player) {
 
         return player;
+    }
+
+    /**
+     * TODO rename this back to #getInstance
+     * @return {@link ServerProxy} instance
+     */
+    public static ServerProxy getInstance2() {
+
+        return ServerProxy.ServerProxyHolder.INSTANCE;
+    }
+
+    /**
+     * instance holder class for lazy and thread-safe initialization
+     */
+    private static class ServerProxyHolder {
+
+        private static final ServerProxy INSTANCE = new ServerProxy();
+
     }
 
 }